ci: Properly support consecutive pre releases
Consecutive pre-releases are not currently possible. Running two release pipelines
with the same pre release flavor would result in the following releases:
1.0.1-rc.0 and 1.0.2-rc.0. But what we want when we run the second pre-release
pipeline is 1.0.1-rc.1.

Now, parse the latest release name and if the latest is a pre-release and its flavor
matches the inputted flavor, we only increment the pre-release version as opposed to
a full version bump.

[DEVX-3004]